[ML] Move field extraction logic to its own package
